Add attributes to generated Python docstrings
Added attributes to the generated Python docstrings. These were absent before, making it hard to tell from an IDE what each class took as inputs. A Google style docstring needed to be used to successful parse these attributes in IDEs, as they're not always directly inside the root class (which reStructuredText doesn't support).

For example, in the generation of Pointset_V1_2_0, the following is added to the docstring (which includes attribute names, types/optional, and descriptions):

    Attributes:
        schema (str)
        locations (Pointset_V1_2_0_Locations): The points in the pointset.
        bounding_box (BoundingBox_V1_0_1): Bounding box of the spatial data.
        coordinate_reference_system (Crs_V1_0_1): Coordinate system of the spatial data
        name (str): Name of the object.
        uuid (BaseObjectProperties_V1_0_1_Uuid): Identifier of the object.
        description (str, optional): Optional field for adding additional description to uniquely identify this object.
        extensions (dict[str, typing.Any], optional): Extended properties that may be associated to the object, but not specified in the schema
        tags (dict[str, str], optional): Key-value pairs of user-defined metadata